+++ database/migrations/2026_03_13_153759_create_invoice_approvals_table.php
@@ -10,7 +10,7 @@ return new class extends Migration
     {
         Schema::create('invoice_approvals', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('invoice_id');
+            $table->integer('invoice_id');
             $table->string('inv_id');
             $table->unsignedBigInteger('requested_by');
             $table->string('requested_by_name');
